requirements.txt
git+https://github.com/Pycord-Development/pycord
humor-langs
translate
PyNaCl
asyncio
silly
utils
httpx
aiofiles
spotipy
psutil
progress
pytimeparse
PyDictionary
lyricsgenius
Pillow
google
wikipedia
alexflipnote.py
speedtest-cli
beautifulsoup4
asyncpraw
akinator.py
udpy
ffmpy
topggpy
pyimgur
rich
discord-together
names
dateparser
qrcode
jishaku
lavalink
